Bug #1596638 pclint告警数2个,涉及文件2个(1)
[root cause  ] PCLint warning modification
[changes     ] modify AE ctrl code
[side effects] no
[self test   ]: AE ctrl source code reviewer
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: libcamera
[test case]: camera test normally
[reviewers   ] jiawei.he
[change_type ] feature_bugfix
[tag_product ] specific

+++ ispalg/smart/smart_ctrl.c
@@ -1526,7 +1526,7 @@ cmr_int _get_atm_curve_v1(cmr_handle *handle,
 			memcpy(uPrevGamma[1], ATMOutput.uGamma, sizeof(uPrevGamma[1]));
 		}
 
-		unsigned int weight = 0.78 * SENSOR_GAMMA_POINT_NUM + 0.5;
+		unsigned int weight = (unsigned int)(0.78 * SENSOR_GAMMA_POINT_NUM + 0.5);
 		isp_atm_smooth(weight, SENSOR_GAMMA_POINT_NUM - 1, uPrevGamma[1], uOutGamma[1], uOutGamma[1]);
 		memcpy(uPrevGamma[1], uOutGamma[1], sizeof(short)*SENSOR_GAMMA_POINT_NUM);
 		if (bATMDump == true) {
